pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deskgpt"
version = "1.0.0"
description = "AI-powered desktop automation tool using OpenAI and Playwright"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "DeskGPT Team"},
]
keywords = ["llm", "automation", "web-browsing", "ai", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.12.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "colorama>=0.4.6",
    "rich>=13.7.0",
]

[project.urls]
Homepage = "https://github.com/your-org/deskgpt"
Repository = "https://github.com/your-org/deskgpt.git"
Issues = "https://github.com/your-org/deskgpt/issues"

[project.scripts]
deskgpt = "deskgpt.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deskgpt*"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88